refactor: renamed types.ts & constants.ts to index.ts
